qwen_edit_lora_video/requirements.txt
replicate>=0.25.0
Pillow>=10.0.0

